0.1.8.7 - Follow module was fixed - Follow Events list now with photos - Docs was updated.

+++ app/Http/Controllers/EventsController.php
@@ -30,6 +30,7 @@ class EventsController extends Controller {
         //return $this->helpReturn(Event::find(2666));
         return $this->helpReturn(Event::with('photos')->findorfail($id));
     }
+
     /**
      * @api {get} /v1/users/events/follow getFollowEventsByUser
      * @apiVersion 0.1.0
@@ -79,6 +80,18 @@ class EventsController extends Controller {
         return $this->helpReturn(User::findorfail($id)->events);
     }
 
+
+    protected function deleteDuplicateInFollows(){
+        $exist = [];
+        foreach(Event_follow::all() as $follow){
+            if(in_array($follow->event_id.':'.$follow->user_id,$exist)){
+                $follow->delete();
+            }else{
+                $exist [] = $follow->event_id.':'.$follow->user_id;
+            }
+        }
+        return;
+    }
     /**
      * @api {post} /v1/users/events/follow followEvents
      * @apiVersion 0.1.0
@@ -91,13 +104,18 @@ class EventsController extends Controller {
      *
      */
     public function follow(Request $request) {
+
         $valid = Validator($request->all(), ['event_id' => 'required']);
         if(!$valid->fails()) {
-            $follow = new Event_follow;
-            $follow->user_id = $request->user->id;
-            $follow->event_id = $request->event_id;
-            $follow->save();
-            return $this->helpInfo($follow);
+            if(!Event_follow::where('user_id', $request->user->id)->where('event_id')->first()) {
+                $follow = new Event_follow;
+                $follow->user_id = $request->user->id;
+                $follow->event_id = $request->event_id;
+                $follow->save();
+                return $this->helpInfo($follow);
+            }else{
+                return $this->helpInfo('Record already exists');
+            }
         } else {
             return $this->helpError('valid', $valid);
         }
@@ -126,7 +144,7 @@ class EventsController extends Controller {
     }
 
     /**
-     * @api {post} /v1/events/unfollow unfollowEvents
+     * @api {post} /v1/users/events/unfollow unfollowEvents
      * @apiVersion 0.1.0
      * @apiName unfollowEvents
      * @apiGroup Events
@@ -139,8 +157,7 @@ class EventsController extends Controller {
     public function unfollow(Request $request) {
         $valid = Validator($request->all(), ['event_id' => 'required']);
         if(!$valid->fails()) {
-            $follow = Event_follow::findorfail($request->event_id);
-            $follow->delete();
+            $follow = Event_follow::where('event_id',$request->event_id)->where('user_id',$request->user->id)->delete();
             return $this->helpInfo();
         } else {
             return $this->helpError('valid', $valid);
@@ -175,19 +192,20 @@ class EventsController extends Controller {
      *
      */
     public function store_save(Request $request) {
-        $rules = ['video'       => false,
-                  'user_id'     => false,
-                  'date_stop'   => false,
-                  'address'     => false,
-                  'place_id'    => false,
-                  'category_id' => 'required',
-                  'title'       => 'required',
-                  'description' => 'required',
-                  'date'        => 'required',
-                  'time'        => 'required',
-                  'type'        => 'required',
-                  'price'       => 'required',
-                  'images'      => false,
+        $rules = [
+            'video'       => false,
+            'user_id'     => false,
+            'date_stop'   => false,
+            'address'     => false,
+            'place_id'    => false,
+            'category_id' => 'required',
+            'title'       => 'required',
+            'description' => 'required',
+            'date'        => 'required',
+            'time'        => 'required',
+            'type'        => 'required',
+            'price'       => 'required',
+            'images'      => false,
         ];
         $category = Category::findorfail($request->category_id);
         if($request->user->balance > $category->post_price) {
@@ -198,8 +216,8 @@ class EventsController extends Controller {
             if(get_class($event) == 'App\Event') {
                 $request->user->save();
                 $info = [];
-                foreach(User::all() as $user){
-                    $info[] = $this->sendPushToUser($user,[
+                foreach(User::all() as $user) {
+                    $info[] = $this->sendPushToUser($user, [
                         'id'    => $event->id,
                         'title' => $event->title,
                         'body'  => $event->description,
@@ -266,19 +284,20 @@ class EventsController extends Controller {
      *
      */
     public function update_save(Request $request, $id) {
-        $rules = ['video'       => false,
-                  'user_id'     => false,
-                  'date_stop'   => false,
-                  'address'     => false,
-                  'place_id'    => false,
-                  'category_id' => 'required',
-                  'title'       => 'required',
-                  'description' => 'required',
-                  'date'        => 'required',
-                  'time'        => 'required',
-                  'type'        => 'required',
-                  'price'       => 'required',
-                  'images'      => false,
+        $rules = [
+            'video'       => false,
+            'user_id'     => false,
+            'date_stop'   => false,
+            'address'     => false,
+            'place_id'    => false,
+            'category_id' => 'required',
+            'title'       => 'required',
+            'description' => 'required',
+            'date'        => 'required',
+            'time'        => 'required',
+            'type'        => 'required',
+            'price'       => 'required',
+            'images'      => false,
         ];
         $request->user_id = $request->user->id;
         $event = $this->fromPostToModel($rules, Event::findorfail($id), $request, 'model');
